scrapper/requirements.txt
httpx
pandas
streamlit
selectolax
mysql-connector-python
